[Process] Add ProcessInterface
